+++ CHANGELOG.md
@@ -56,6 +56,8 @@ Main (unreleased)
 
 - Propagate request metadata for `faro.receiver` to downstream components. (@hainenber)
 
+- Configurable default `instance` label for logs scraped by `loki.source.kubernetes`. (@hainenber)
+
 ### Features
 
 - A new `loki.rules.kubernetes` component that discovers `PrometheusRule` Kubernetes resources and loads them into a Loki Ruler instance. (@EStork09)